fixed test script

+++ testScript.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 for TESTCASE in {1..21}
 do
-  echo "=====TESTING $TESTCASE.micro========"　
-  ./Micro testcases/input/test$TESTCASE.micro > out
-  cat testcases/output/test$TESTCASE.out > out2
-  diff out out2
+  if [ -e testcases/input/test$TESTCASE.micro ]; then
+    echo "=====TESTING $TESTCASE.micro========"　
+    ./Micro testcases/input/test$TESTCASE.micro > out
+    cat testcases/output/test$TESTCASE.out > out2
+    diff out out2
+  fi
+  if [ -e out ]; then
+    rm out
+    rm out2
+  fi
 done
